+++ lib/src/entity/entity_field.dart
@@ -11,8 +11,8 @@ class Field<T> extends FieldBase<T> {
   })  : assert(T is! List),
         assert(T is! Entity || defaultValue == null),
         super(entity, name: name, label: label) {
-    final _value = defaultValue ?? T.getDefault();
-    if (_value != null) _createDefault = () => _value;
+    final val = defaultValue ?? T.getDefault();
+    if (val != null) _createDefault = () => val;
   }
 
   T Function()? _createDefault;
@@ -25,9 +25,9 @@ class Field<T> extends FieldBase<T> {
   T? get value {
     _getDefault() {
       if (isLoaded) return null;
-      final _value = innerDefaultValue();
-      entity.data[name] = _value;
-      return _value;
+      final val = innerDefaultValue();
+      entity.data[name] = val;
+      return val;
     }
 
     return entity[name] ?? _getDefault();
@@ -45,8 +45,8 @@ class Field<T> extends FieldBase<T> {
   /// In order to assign null value, use [value] instead.
   T call([T? value]) {
     if (value != null) this.value = value;
-    final _value = this.value;
-    assert(_value != null, '${entity.runtimeType}.$name() has a null value!');
+    final val = this.value;
+    assert(val != null, '${entity.runtimeType}.$name() has a null value!');
     return this.value!;
   }
 
@@ -84,9 +84,11 @@ extension FieldExtensions<T> on Field<T> {
 extension FieldOfEntityExtensions<E extends Entity> on Field<E> {
   Field<E> register(
     EntityBuilder<E> createEntity, {
-    bool auto = false,
+    bool defaultInstance = false,
   }) {
-    if (auto) _createDefault = () => createEntity().._parentRef = this;
+    if (defaultInstance) {
+      _createDefault = () => createEntity().._parentRef = this;
+    }
     _fieldOnLoading = (rawData) {
       final instance = value ?? createEntity()
         .._parentRef = this;