+++ lib/src/services/services.dart
@@ -1,4 +1,6 @@
 // https://github.com/FilledStacks/stacked
 export 'dialog/dialog_service.dart';
 export 'launcher/launcher.dart';
+export 'models/overlay_request.dart';
+export 'models/overlay_response.dart';
 export 'snackbar/snackbar_service.dart';