+++ lib/src/utils/utils.dart
@@ -1,4 +1,4 @@
-export 'src/conditional_switch.dart';
 export 'src/cross_file_image.dart';
 export 'src/crypto.dart';
+export 'src/debug.dart';
 export 'src/uuid.dart';