+++ publish_dryrun.sh
@@ -1,0 +1,5 @@
+flutter clean
+
+flutter pub get
+
+flutter pub publish --dry-run